Decouple paginated collection from Hydra collections
