+++ Resources/config/json_ld.xml
@@ -26,6 +26,7 @@
             <argument type="service" id="api.mapping.class_metadata_factory" />
             <argument type="service" id="api.json_ld.context_builder" />
             <argument type="service" id="property_accessor" />
+            <argument type="service" id="api.resource_resolver" />
             <argument type="service" id="api.name_converter" on-invalid="ignore" />
 
             <tag name="serializer.normalizer" />
